+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## v0.3.14 - Request DB Switching 
+  * Commit: 2026-04-14 08:30
+  * Request.AddDB
+  * Request.SwitchDB
 ## v0.3.13 - Change Query Conditions
   * Commit: 2026-04-13 13:57
   * Rename DualConditions